requirements.txt
Flask>=2.0.0,<3.0.0
Flask-CORS>=3.0.0,<4.0.0
python-dotenv>=0.19.0,<1.0.0
Werkzeug>=2.0.0,<3.0.0